Contig stitcher: factor out context into a separate file
